[bug] fix order of insertion of decrypted docs
This commit actually does some different things:

  * When doing asynchronous decryption of incoming documents in soledad client
    during a sync, there was the possibility that a document corresponding to
    a newer generation would be decrypted and inserted in the local database
    before a document corresponding to an older generation. When this
    happened, the metadata about the target database (i.e. its locally-known
    generation) would be first updated to the newer generation, and then an
    attempt to insert a document corresponding to an older generation would
    cause the infamous InvalidGeneration error.

    To fix that we use the sync-index information that is contained in the
    sync stream to correctly find the insertable docs to be inserted in the
    local database, thus avoiding the problem described above.

  * Refactor the sync encrypt/decrypt pool to its own file.

  * Fix the use of twisted adbapi with multiprocessing.

Closes: #6757.
