+++ client/changes/bug_6757_fix-order-of-insertion-when-syncing
@@ -1,0 +1,2 @@
+  o Fix the order of insertion of documents when using workers for decrypting
+    incoming documents during a sync. Closes #6757.
